backend/requirements.txt
paho-mqtt==1.6.1
numpy==1.21.0
scipy==1.7.0
matplotlib==3.4.0
